+++ tests/testthat/test-counts.R
@@ -17,31 +17,31 @@ test_that("counts are computed correctly", {
         full.names = TRUE)
     d <- readr::read_csv(result_file)
     d$mean <- round(d$mean)
-    d <- d[order(d$spatial_code),]
+    d <- d[order(d$location_code),]
 
     # check results
     # reference values have been calculated by using spread sheet
     # 'testthat-manual-beach-litter-nl-2012-2017.ods'
     expect_equal(
-        d[d$type_name == "TC", "mean"][[1]],
-        c(377, 347, 299, 231)
+        d[d$`type/group_name` == "TC", "mean"][[1]],
+        c(377, 347, 231, 299)
     )
     expect_equal(
-        d[d$type_name == "SUP", "mean"][[1]],
-        c(91, 82, 61, 44)
+        d[d$`type/group_name` == "SUP", "mean"][[1]],
+        c(91, 82, 44, 61)
     )
     expect_equal(
-        d[d$type_name == "FISH", "mean"][[1]],
-        c(162, 143, 139, 106)
+        d[d$`type/group_name` == "FISH", "mean"][[1]],
+        c(162, 143, 106, 139)
     )
     expect_equal(
-        d[d$type_name == "PLASTIC", "mean"][[1]],
-        c(341, 312, 265, 210)
+        d[d$`type/group_name` == "PLASTIC", "mean"][[1]],
+        c(341, 312, 210, 265)
     )
 
-    # replace spatial codes by a single spatial code
+    # replace location_codes by a single location_code
     d <- readr::read_csv(file.path(tmp_dir, "beach-litter-nl-2012-2017.csv"))
-    d$spatial_code <- "The Netherlands"
+    d$location_code <- "The Netherlands"
     readr::write_csv(d, file.path(tmp_dir, "beach-litter-nl-2012-2017.csv"))
     
     # run litteR
@@ -59,19 +59,19 @@ test_that("counts are computed correctly", {
     # reference values have been calculated by using spread sheet
     # 'testthat-manual-beach-litter-nl-2012-2017.ods'
     expect_equal(
-        d[d$type_name == "TC", "mean"][[1]],
+        d[d$`type/group_name` == "TC", "mean"][[1]],
         314
     )
     expect_equal(
-        d[d$type_name == "SUP", "mean"][[1]],
+        d[d$`type/group_name` == "SUP", "mean"][[1]],
         70
     )
     expect_equal(
-        d[d$type_name == "FISH", "mean"][[1]],
+        d[d$`type/group_name` == "FISH", "mean"][[1]],
         138
     )
     expect_equal(
-        d[d$type_name == "PLASTIC", "mean"][[1]],
+        d[d$`type/group_name` == "PLASTIC", "mean"][[1]],
         282
     )
 
@@ -101,19 +101,19 @@ test_that("counts are computed correctly", {
     # reference values have been calculated by using spread sheet
     # 'testthat-manual-beach-litter-nl-2012-2017.ods'
     expect_equal(
-        d[d$type_name == "TC", "mean"][[1]],
+        d[d$`type/group_name` == "TC", "mean"][[1]],
         314
     )
     expect_equal(
-        d[d$type_name == "SUP", "mean"][[1]],
+        d[d$`type/group_name` == "SUP", "mean"][[1]],
         70
     )
     expect_equal(
-        d[d$type_name == "FISH", "mean"][[1]],
+        d[d$`type/group_name` == "FISH", "mean"][[1]],
         138
     )
     expect_equal(
-        d[d$type_name == "PLASTIC", "mean"][[1]],
+        d[d$`type/group_name` == "PLASTIC", "mean"][[1]],
         282
     )
 })